+++ lib/geordi/commands/util.rb
